setPending needs to account for the new filtered counter
when the filtered counter was added I failed to account for it
in setPending and resetCounters. This caused the pending count to
not be calculated correctly, it was exactly filtered count too large.

In the future it may be better to use a single counter for pending
that is incremented on 'request' and decremeted on 'complete'