Adds reqInfo object to pool.request callback
This change is a little specific, so no pressure to merge. It adds an additional callback argument
to pool.request, namely, requestInfo. It is a dictionary containing information about the node that
responded to the request, nodes that failed to respond to the request, and the number of total
retries.

It does this by creating a closure when preparing to handle the response containing the node that
the response is coming from.